refactor TlsUtils class
