+++ pinot-common/src/main/java/org/apache/pinot/common/utils/tls/TlsUtils.java
@@ -18,32 +18,19 @@
  */
 package org.apache.pinot.common.utils.tls;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import io.netty.handler.ssl.ClientAuth;
 import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslContextBuilder;
 import io.netty.handler.ssl.SslProvider;
 import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
-import java.io.IOException;
 import java.io.InputStream;
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.nio.file.FileSystems;
-import java.nio.file.Path;
-import java.nio.file.StandardWatchEventKinds;
-import java.nio.file.WatchEvent;
-import java.nio.file.WatchKey;
-import java.nio.file.WatchService;
 import java.security.KeyStore;
 import java.security.SecureRandom;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.Executors;
 import java.util.concurrent.atomic.AtomicReference;
 import javax.net.ssl.HttpsURLConnection;
 import javax.net.ssl.KeyManagerFactory;
@@ -51,14 +38,10 @@ import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManagerFactory;
 import nl.altindag.ssl.SSLFactory;
 import nl.altindag.ssl.exception.GenericSSLContextException;
-import nl.altindag.ssl.keymanager.HotSwappableX509ExtendedKeyManager;
-import nl.altindag.ssl.trustmanager.HotSwappableX509ExtendedTrustManager;
-import nl.altindag.ssl.util.SSLFactoryUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.http.ssl.SSLContexts;
 import org.apache.pinot.common.config.TlsConfig;
 import org.apache.pinot.spi.env.PinotConfiguration;
-import org.apache.pinot.spi.utils.retry.RetryPolicies;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -244,13 +227,13 @@ public final class TlsUtils {
       String trustStoreType, String trustStorePath, String trustStorePassword) {
     try {
       SecureRandom secureRandom = new SecureRandom();
-      SSLFactory sslFactory = createSSLFactory(keyStoreType, keyStorePath, keyStorePassword,
+      SSLFactory sslFactory = RenewableTlsUtils.createSSLFactory(keyStoreType, keyStorePath, keyStorePassword,
           trustStoreType, trustStorePath, trustStorePassword,
           "SSL", secureRandom, true, false);
       if (isKeyOrTrustStorePathNullOrHasFileScheme(keyStorePath)
           && isKeyOrTrustStorePathNullOrHasFileScheme(trustStorePath)) {
-        enableAutoRenewalFromFileStoreForSSLFactory(sslFactory, keyStoreType, keyStorePath, keyStorePassword,
-            trustStoreType, trustStorePath, trustStorePassword, "SSL", secureRandom, false);
+        RenewableTlsUtils.enableAutoRenewalFromFileStoreForSSLFactory(sslFactory, keyStoreType, keyStorePath,
+            keyStorePassword, trustStoreType, trustStorePath, trustStorePassword, "SSL", secureRandom, false);
       }
       // HttpsURLConnection
       HttpsURLConnection.setDefaultSSLSocketFactory(sslFactory.getSslSocketFactory());
@@ -317,7 +300,7 @@ public final class TlsUtils {
    * @param tlsConfig TLS config
    */
   public static SslContext buildClientContext(TlsConfig tlsConfig) {
-    SSLFactory sslFactory = createSSLFactoryAndEnableAutoRenewalWhenUsingFileStores(tlsConfig);
+    SSLFactory sslFactory = RenewableTlsUtils.createSSLFactoryAndEnableAutoRenewalWhenUsingFileStores(tlsConfig);
     SslContextBuilder sslContextBuilder =
         SslContextBuilder.forClient().sslProvider(SslProvider.valueOf(tlsConfig.getSslProvider()));
     sslFactory.getKeyManagerFactory().ifPresent(sslContextBuilder::keyManager);
@@ -338,7 +321,7 @@ public final class TlsUtils {
     if (tlsConfig.getKeyStorePath() == null) {
       throw new IllegalArgumentException("Must provide key store path for secured server");
     }
-    SSLFactory sslFactory = createSSLFactoryAndEnableAutoRenewalWhenUsingFileStores(tlsConfig);
+    SSLFactory sslFactory = RenewableTlsUtils.createSSLFactoryAndEnableAutoRenewalWhenUsingFileStores(tlsConfig);
     SslContextBuilder sslContextBuilder = SslContextBuilder.forServer(sslFactory.getKeyManagerFactory().get())
         .sslProvider(SslProvider.valueOf(tlsConfig.getSslProvider()));
     sslFactory.getTrustManagerFactory().ifPresent(sslContextBuilder::trustManager);
@@ -357,220 +340,12 @@ public final class TlsUtils {
    *
    * @param keyOrTrustStorePath key store or trust store path in String format.
    */
-  public static boolean isKeyOrTrustStorePathNullOrHasFileScheme(String keyOrTrustStorePath) {
+  static boolean isKeyOrTrustStorePathNullOrHasFileScheme(String keyOrTrustStorePath) {
     try {
       return keyOrTrustStorePath == null
           || makeKeyOrTrustStoreUrl(keyOrTrustStorePath).toURI().getScheme().startsWith(FILE_SCHEME);
     } catch (Exception e) {
       throw new RuntimeException(e);
-    }
-  }
-
-  /**
-   * Enables auto renewal of SSLFactory when
-   * 1. the {@link SSLFactory} is created with a key manager and trust manager swappable
-   * 2. the key store is null or a local file
-   * 3. the trust store is null or a local file
-   * 4. the key store or trust store file changes.
-   * @param sslFactory the {@link SSLFactory} to enable key manager and trust manager auto renewal
-   * @param tlsConfig the {@link TlsConfig} to get the key store and trust store information
-   */
-  public static void enableAutoRenewalFromFileStoreForSSLFactory(SSLFactory sslFactory, TlsConfig tlsConfig) {
-    enableAutoRenewalFromFileStoreForSSLFactory(sslFactory,
-        tlsConfig.getKeyStoreType(), tlsConfig.getKeyStorePath(), tlsConfig.getKeyStorePassword(),
-        tlsConfig.getTrustStoreType(), tlsConfig.getTrustStorePath(), tlsConfig.getTrustStorePassword(),
-        null, null, tlsConfig.isInsecure());
-  }
-
-  static void enableAutoRenewalFromFileStoreForSSLFactory(SSLFactory sslFactory, String keyStoreType,
-      String keyStorePath, String keyStorePassword, String trustStoreType, String trustStorePath,
-      String trustStorePassword, String sslContextProtocol, SecureRandom secureRandom, boolean isInsecure) {
-    try {
-      URL keyStoreURL = keyStorePath == null ? null : makeKeyOrTrustStoreUrl(keyStorePath);
-      URL trustStoreURL = trustStorePath == null ? null : makeKeyOrTrustStoreUrl(trustStorePath);
-      if (keyStoreURL != null) {
-        Preconditions.checkArgument(
-            keyStoreURL.toURI().getScheme().startsWith(FILE_SCHEME),
-            "key store path must be a local file path or null when SSL auto renew is enabled");
-        Preconditions.checkArgument(
-            sslFactory.getKeyManager().isPresent()
-                && sslFactory.getKeyManager().get() instanceof HotSwappableX509ExtendedKeyManager,
-            "key manager of the existing SSLFactory must be swappable"
-        );
-      }
-      if (trustStoreURL != null) {
-        Preconditions.checkArgument(
-            trustStoreURL.toURI().getScheme().startsWith(FILE_SCHEME),
-            "trust store path must be a local file path or null when SSL auto renew is enabled");
-        Preconditions.checkArgument(
-            sslFactory.getTrustManager().isPresent()
-                && sslFactory.getTrustManager().get() instanceof HotSwappableX509ExtendedTrustManager,
-            "trust manager of the existing SSLFactory must be swappable"
-        );
-      }
-      // The reloadSslFactoryWhenFileStoreChanges is a blocking call, so we need to create a new thread to run it.
-      // Creating a new thread to run the reloadSslFactoryWhenFileStoreChanges is costly; however, unless we
-      // invoke the createAutoRenewedSSLFactoryFromFileStore method crazily, this should not be a problem.
-      Executors.newSingleThreadExecutor().execute(() -> {
-        try {
-          reloadSslFactoryWhenFileStoreChanges(sslFactory,
-              keyStoreType, keyStorePath, keyStorePassword,
-              trustStoreType, trustStorePath, trustStorePassword,
-              sslContextProtocol, secureRandom, isInsecure);
-        } catch (Exception e) {
-          throw new RuntimeException(e);
-        }
-      });
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  @VisibleForTesting
-  static void reloadSslFactoryWhenFileStoreChanges(SSLFactory baseSslFactory,
-      String keyStoreType, String keyStorePath, String keyStorePassword,
-      String trustStoreType, String trustStorePath, String trustStorePassword,
-      String sslContextProtocol, SecureRandom secureRandom, boolean isInsecure)
-      throws IOException, URISyntaxException, InterruptedException {
-    LOGGER.info("Enable auto renewal of SSLFactory {} when key store {} or trust store {} changes",
-        baseSslFactory, keyStorePath, trustStorePath);
-    WatchService watchService = FileSystems.getDefault().newWatchService();
-    Map<WatchKey, Set<Path>> watchKeyPathMap = new HashMap<>();
-    registerFile(watchService, watchKeyPathMap, keyStorePath);
-    registerFile(watchService, watchKeyPathMap, trustStorePath);
-    int maxSslFactoryReloadingAttempts = 3;
-    int sslFactoryReloadingRetryDelayMs = 1000;
-    WatchKey key;
-    while ((key = watchService.take()) != null) {
-      for (WatchEvent<?> event : key.pollEvents()) {
-        Path changedFile = (Path) event.context();
-        if (watchKeyPathMap.get(key).contains(changedFile)) {
-          LOGGER.info("Detected change in file: {}, try to renew SSLFactory {} "
-              + "(built from key store {} and truststore {})",
-              changedFile, baseSslFactory, keyStorePath, trustStorePath);
-          try {
-            // Need to retry a few times because when one file (key store or trust store) is updated, the other file
-            // (trust store or key store) may not have been fully written yet, so we need to wait a bit and retry.
-            RetryPolicies.fixedDelayRetryPolicy(maxSslFactoryReloadingAttempts, sslFactoryReloadingRetryDelayMs)
-                .attempt(() -> {
-                  try {
-                    SSLFactory updatedSslFactory =
-                        createSSLFactory(keyStoreType, keyStorePath, keyStorePassword, trustStoreType, trustStorePath,
-                            trustStorePassword, sslContextProtocol, secureRandom, false, isInsecure);
-                    SSLFactoryUtils.reload(baseSslFactory, updatedSslFactory);
-                    LOGGER.info("Successfully renewed SSLFactory {} (built from key store {} and truststore {}) on file"
-                        + " {} changes", baseSslFactory, keyStorePath, trustStorePath, changedFile);
-                    return true;
-                  } catch (Exception e) {
-                    LOGGER.info(
-                        "Encountered issues when renewing SSLFactory {} (built from key store {} and truststore {}) on "
-                            + "file {} changes", baseSslFactory, keyStorePath, trustStorePath, changedFile, e);
-                    return false;
-                  }
-                });
-          } catch (Exception e) {
-            LOGGER.error(
-                "Failed to renew SSLFactory {} (built from key store {} and truststore {}) on file {} changes after {} "
-                    + "retries", baseSslFactory, keyStorePath, trustStorePath, changedFile,
-                maxSslFactoryReloadingAttempts, e);
-          }
-        }
-      }
-      key.reset();
-    }
-  }
-
-  @VisibleForTesting
-  static void registerFile(WatchService watchService, Map<WatchKey, Set<Path>> keyPathMap, String filePath)
-      throws IOException, URISyntaxException {
-    if (filePath == null) {
-      return;
-    }
-    Path path = Path.of(makeKeyOrTrustStoreUrl(filePath).getPath());
-    WatchKey key = path.getParent().register(watchService, StandardWatchEventKinds.ENTRY_MODIFY);
-    keyPathMap.computeIfAbsent(key, k -> new HashSet<>());
-    keyPathMap.get(key).add(path.getFileName());
-  }
-
-  /**
-   * Create a {@link SSLFactory} instance with identity material and trust material swappable for a given TlsConfig,
-   * and nables auto renewal of the {@link SSLFactory} instance when
-   * 1. the {@link SSLFactory} is created with a key manager and trust manager swappable
-   * 2. the key store is null or a local file
-   * 3. the trust store is null or a local file
-   * 4. the key store or trust store file changes.
-   * @param tlsConfig {@link TlsConfig}
-   * @return a {@link SSLFactory} instance with identity material and trust material swappable
-   */
-  public static SSLFactory createSSLFactoryAndEnableAutoRenewalWhenUsingFileStores(TlsConfig tlsConfig) {
-    SSLFactory sslFactory = createSSLFactory(tlsConfig);
-    if (isKeyOrTrustStorePathNullOrHasFileScheme(tlsConfig.getKeyStorePath())
-        && isKeyOrTrustStorePathNullOrHasFileScheme(tlsConfig.getTrustStorePath())) {
-      enableAutoRenewalFromFileStoreForSSLFactory(sslFactory, tlsConfig);
-    }
-    return sslFactory;
-  }
-
-  /**
-   * Create a {@link SSLFactory} instance with identity material and trust material swappable for a given TlsConfig
-   * @param tlsConfig {@link TlsConfig}
-   * @return a {@link SSLFactory} instance with identity material and trust material swappable
-   */
-  public static SSLFactory createSSLFactory(TlsConfig tlsConfig) {
-    return createSSLFactory(
-        tlsConfig.getKeyStoreType(), tlsConfig.getKeyStorePath(), tlsConfig.getKeyStorePassword(),
-        tlsConfig.getTrustStoreType(), tlsConfig.getTrustStorePath(), tlsConfig.getTrustStorePassword(),
-        null, null, true, tlsConfig.isInsecure());
-  }
-
-  static SSLFactory createSSLFactory(
-      String keyStoreType, String keyStorePath, String keyStorePassword,
-      String trustStoreType, String trustStorePath, String trustStorePassword,
-      String sslContextProtocol, SecureRandom secureRandom, boolean keyAndTrustMaterialSwappable, boolean isInsecure) {
-    try {
-      SSLFactory.Builder sslFactoryBuilder = SSLFactory.builder();
-      InputStream keyStoreStream = null;
-      InputStream trustStoreStream = null;
-      if (keyStorePath != null) {
-        Preconditions.checkNotNull(keyStorePassword, "key store password must not be null");
-        keyStoreStream = makeKeyOrTrustStoreUrl(keyStorePath).openStream();
-        if (keyAndTrustMaterialSwappable) {
-          sslFactoryBuilder.withSwappableIdentityMaterial();
-        }
-        sslFactoryBuilder.withIdentityMaterial(keyStoreStream, keyStorePassword.toCharArray(), keyStoreType);
-      }
-      if (isInsecure) {
-        if (keyAndTrustMaterialSwappable) {
-          sslFactoryBuilder.withSwappableTrustMaterial();
-        }
-        sslFactoryBuilder.withUnsafeTrustMaterial();
-      } else if (trustStorePath != null) {
-        Preconditions.checkNotNull(trustStorePassword, "trust store password must not be null");
-        trustStoreStream = makeKeyOrTrustStoreUrl(trustStorePath).openStream();
-        if (keyAndTrustMaterialSwappable) {
-          sslFactoryBuilder.withSwappableTrustMaterial();
-        }
-        sslFactoryBuilder.withTrustMaterial(trustStoreStream, trustStorePassword.toCharArray(), trustStoreType);
-      }
-      if (sslContextProtocol != null) {
-        sslFactoryBuilder.withSslContextAlgorithm(sslContextProtocol);
-      }
-      if (secureRandom != null) {
-        sslFactoryBuilder.withSecureRandom(secureRandom);
-      }
-      SSLFactory sslFactory = sslFactoryBuilder.build();
-      if (keyStoreStream != null) {
-        keyStoreStream.close();
-      }
-      if (trustStoreStream != null) {
-        trustStoreStream.close();
-      }
-      LOGGER.info("Successfully created SSLFactory {} with key store {} and trust store {}. "
-              + "Key and trust material swappable: {}",
-          sslFactory, keyStorePath, trustStorePath, keyAndTrustMaterialSwappable);
-      return sslFactory;
-    } catch (Exception e) {
-      throw new IllegalStateException(e);
     }
   }
 }
